Refactors Docker Compose network configuration
Updates the Docker Compose configuration to define the `pourymovie` network internally instead of relying on an external one.

This change simplifies the deployment process by ensuring that the network is created automatically when the containers are started, and removes the need for manual network creation.

Also, re-enables the frontend runner in the runner docker-compose.